requirements.txt
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
pgvector>=0.2.0
langchain-text-splitters>=0.3.0
sentence-transformers>=3.0.0
